fix(LetMeShip): Move api to doctype
